rename NoticeBar -> ContentUnavailableView
