+++ Screenbox/App.xaml
@@ -149,7 +149,7 @@
                     </DataTemplate>
                 </ResourceDictionary>
 
-                <ResourceDictionary Source="ms-appx:///Controls/NoticeBar/NoticeBar.xaml" />
+                <ResourceDictionary Source="ms-appx:///Controls/ContentUnavailableView/ContentUnavailableView.xaml" />
                 <ResourceDictionary Source="ms-appx:///Controls/TitleBar/TitleBar.xaml" />
                 <ResourceDictionary Source="ms-appx:///Styles/Button.xaml" />
                 <ResourceDictionary Source="ms-appx:///Styles/Icon.xaml" />
