+++ Screenbox/Pages/FolderViewPage.xaml
@@ -156,7 +156,7 @@
         </GridView>
 
         <!--  Empty content  -->
-        <controls:NoticeBar
+        <controls:ContentUnavailableView
             x:Name="NoContentPanel"
             Grid.Row="2"
             Margin="{StaticResource TopLargeMargin}"
