+++ Screenbox/Pages/VideosPage.xaml
@@ -123,7 +123,7 @@
                     HorizontalScrollMode="Disabled"
                     VerticalScrollBarVisibility="Auto"
                     Visibility="Collapsed">
-                    <controls:NoticeBar
+                    <controls:ContentUnavailableView
                         x:Name="NoVideosPanel"
                         Title="{x:Bind strings:Resources.NoVideosPanelHeader}"
                         IsOpen="False"
@@ -135,7 +135,7 @@
                             FontSize="{StaticResource NoContentPanelIconFontSize}"
                             Foreground="{ThemeResource AccentTextFillColorPrimaryBrush}"
                             Glyph="{StaticResource MoviesOffGlyph}" />
-                        <controls:NoticeBar.ActionContent>
+                        <controls:ContentUnavailableView.ActionContent>
                             <Button
                                 AccessKey="{x:Bind strings:KeyboardResources.CommandAddFolderKey}"
                                 AutomationProperties.Name="{x:Bind strings:Resources.AddFolder}"
@@ -152,8 +152,8 @@
                                     <TextBlock Margin="8,0,0,0" Text="{x:Bind strings:Resources.AddFolder}" />
                                 </StackPanel>
                             </Button>
-                        </controls:NoticeBar.ActionContent>
-                    </controls:NoticeBar>
+                        </controls:ContentUnavailableView.ActionContent>
+                    </controls:ContentUnavailableView>
                 </ScrollViewer>
             </Grid>
         </muxc:NavigationView>
